Move schema evolution tests to dedicated traits
