+++ spark/src/test/scala/org/apache/spark/sql/delta/MergeIntoNotMatchedBySourceSuite.scala
@@ -527,75 +527,4 @@ trait MergeIntoNotMatchedBySourceSuite extends MergeIntoSuiteBase {
       clauses = updateNotMatched(set = "id = t.id * 10"))
     checkAnswer(readDeltaTable(target), Seq(0, 10, 2, 30, 4, 50, 6, 70, 8, 90).toDF("id"))
   }
-
-  // Test schema evolution with NOT MATCHED BY SOURCE clauses.
-  testEvolution("new column with insert * and delete not matched by source")(
-    sourceData = Seq((1, 1, "extra1"), (2, 2, "extra2")).toDF("key", "value", "extra"),
-    targetData = Seq((0, 0), (1, 10), (3, 30)).toDF("key", "value"),
-    clauses = insert("*") ::
-      deleteNotMatched() :: Nil,
-    expected = Seq(
-      // (0, 0) Not matched by source, deleted
-      (1, 10, null), // Matched, updated
-      (2, 2, "extra2") // Not matched by target, inserted
-      // (3, 30) Not matched by source, deleted
-    ).toDF("key", "value", "extra"),
-    expectedWithoutEvolution = Seq((1, 10), (2, 2)).toDF("key", "value"))
-
-  testEvolution("new column with insert * and conditional update not matched by source")(
-    targetData = Seq((0, 0), (1, 10), (3, 30)).toDF("key", "value"),
-    sourceData = Seq((1, 1, "extra1"), (2, 2, "extra2")).toDF("key", "value", "extra"),
-    clauses = insert("*") ::
-      updateNotMatched(condition = "key > 0", set = "value = value + 1") :: Nil,
-    expected = Seq(
-      (0, 0, null), // Not matched by source, no change
-      (1, 10, null), // Matched, no change
-      (2, 2, "extra2"), // Not matched by target, inserted
-      (3, 31, null) // Not matched by source, updated
-    ).toDF("key", "value", "extra"),
-    expectedWithoutEvolution = Seq((0, 0), (1, 10), (2, 2), (3, 31)).toDF("key", "value"))
-
-  testEvolution("new column not inserted and conditional update not matched by source")(
-    targetData = Seq((0, 0), (1, 10), (3, 30)).toDF("key", "value"),
-    sourceData = Seq((1, 1, "extra1"), (2, 2, "extra2")).toDF("key", "value", "extra"),
-    clauses = updateNotMatched(condition = "key > 0", set = "value = value + 1") :: Nil,
-    expected = Seq(
-      (0, 0), // Not matched by source, no change
-      (1, 10), // Matched, no change
-      (3, 31) // Not matched by source, updated
-    ).toDF("key", "value"),
-    expectedWithoutEvolution = Seq((0, 0), (1, 10), (3, 31)).toDF("key", "value"))
-
-  testEvolution("new column referenced in matched condition but not inserted")(
-    targetData = Seq((0, 0), (1, 10), (3, 30)).toDF("key", "value"),
-    sourceData = Seq((1, 1, "extra1"), (2, 2, "extra2")).toDF("key", "value", "extra"),
-    clauses = delete(condition = "extra = 'extra1'") ::
-      updateNotMatched(condition = "key > 0", set = "value = value + 1") :: Nil,
-    expected = Seq(
-      (0, 0), // Not matched by source, no change
-      // (1, 10), Matched, deleted
-      (3, 31) // Not matched by source, updated
-    ).toDF("key", "value"),
-    expectedWithoutEvolution = Seq((0, 0), (3, 31)).toDF("key", "value"))
-
-  testEvolution("matched update * and conditional update not matched by source")(
-    targetData = Seq((0, 0), (1, 10), (3, 30)).toDF("key", "value"),
-    sourceData = Seq((1, 1, "extra1"), (2, 2, "extra2")).toDF("key", "value", "extra"),
-    clauses = update("*") ::
-      updateNotMatched(condition = "key > 0", set = "value = value + 1") :: Nil,
-    expected = Seq(
-      (0, 0, null), // Not matched by source, no change
-      (1, 1, "extra1"), // Matched, updated
-      (3, 31, null) // Not matched by source, updated
-    ).toDF("key", "value", "extra"),
-    expectedWithoutEvolution = Seq((0, 0), (1, 1), (3, 31)).toDF("key", "value"))
-
-  // Migrating new column via WHEN NOT MATCHED BY SOURCE is not allowed.
-  testEvolution("update new column with not matched by source fails")(
-    targetData = Seq((0, 0), (1, 10), (3, 30)).toDF("key", "value"),
-    sourceData = Seq((1, 1, "extra3"), (2, 2, "extra2")).toDF("key", "value", "extra"),
-    clauses = updateNotMatched("extra = s.extra") :: Nil,
-    expectErrorContains = "cannot resolve extra in UPDATE clause",
-    expectErrorWithoutEvolutionContains = "cannot resolve extra in UPDATE clause")
-
 }
